requirements.txt
pytest~=8.2.2
Appium-Python-Client~=4.0.0
selenium~=4.22.0
python-dotenv~=1.0.1